Swith to VideoInfo2 structure This allows software deinterlacing both inside the codec as well as interlace detection for use in myth. Also deblocking can now be disabled for large performance gains

+++ trunk/loader/wine/windef.h
@@ -609,10 +609,10 @@ typedef struct tagPOINTS
 /* The RECT structure */
 typedef struct tagRECT
 {
-    short  left;
-    short  top;
-    short  right;
-    short  bottom;
+    LONG left;
+    LONG top;
+    LONG right;
+    LONG bottom;
 } RECT, *PRECT, *LPRECT;
 typedef const RECT *LPCRECT;
